separate ragas and streaming events
